// src/store/modules/student/mutation-types.ts
// TODO: Remove no-shadow eslint disable after fixed: https://github.com/typescript-eslint/typescript-eslint/issues/2484
// eslint-disable-next-line no-shadow
export enum StudentMutationTypes {
  SET_TOKEN = 'SET_STUDENT_TOKEN',
  SET_USER = 'SET_STUDENT',
  SET_BASKET_COUNT = 'SET_BASET_COUNT',
  SET_CURRENT_SHOP_INFO = 'SET_CURRENT_SHOP_INFO',
  SET_CURRENT_SUPPORT_PERSON = 'SET_CURRENT_SUPPORT_PERSON',
  SET_CURRENT_QUESTIONS = 'SET_CURRENT_QUESTIONS',
  SET_CURRENT_QUESTION = 'SET_CURRENT_QUESTION',
  SET_CURRENT_DUEL = 'SET_CURRENT_SET_CURRENT_DUEL',
  SET_CURRENT_SELFTEST_REPORT = 'SET_CURRENT_SELFTEST_REPORT',
  SET_ONE_QUESTION_FOR_SELFTEST = 'SET_ONE_QUESTION_FOR_SELFTEST'
}